requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.3.20
langchain_openai==0.3.7
langchain_core==0.3.72
langchain_text_splitters==0.3.9
python-dotenv==1.0.0
pydantic==2.11.7